+++ include/network/web.h
@@ -25,7 +25,7 @@
 
 #include <filesystem>
 #include <string>
-#include <json/json.h>
+#include <boost/json.hpp>
 #include "curleasy.h"
 
 namespace Nickvision::Network::Web
@@ -39,9 +39,9 @@ namespace Nickvision::Network::Web
     /**
      * @brief Fetches a json string from a url.
      * @param url The url of the json string
-     * @return The fetched json object
+     * @return The fetched json value
      */
-    Json::Value fetchJson(const std::string& url);
+    boost::json::value fetchJson(const std::string& url);
     /**
      * @brief Downloads a file to disk.
      * @param url The url of the file to download