+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2024.9.0 (next)
+### Breaking Changes
+- Replaced `jsoncpp` library with `boost-json`
+#### App
+- `Nickvision::App::DataFileBase`'s `m_json` object is now of type `boost::json::object`
+### New APIs
+None
+### Fixes
+None
+
 ## 2024.8.3
 ### Breaking Changes
 #### Keyring